new mode==s changes

+++ src/PollProcessorFunction/Models/PollRequest.cs
@@ -5,6 +5,9 @@ public sealed class PollRequest
     public string AppId { get; set; } = "";
     public string Mode { get; set; } = "F";
     public bool Reset { get; set; } = true;
-    public int PollForHours { get; set; } = 2;
+    /// <summary>
+    /// How long to poll (hours). Supports fractions, e.g. 0.25 = 15 minutes.
+    /// </summary>
+    public double PollForHours { get; set; } = 2;
 }
